+++ source/ChanSort/GlobalImageCollection.cs
@@ -49,7 +49,12 @@ namespace ChanSort.Ui
       // 
       // sharedImageCollection1
       // 
+      // 
+      // 
+      // 
       this.sharedImageCollection1.ImageSource.ImageStream = ((DevExpress.Utils.ImageCollectionStreamer)(resources.GetObject("sharedImageCollection1.ImageSource.ImageStream")));
+      this.sharedImageCollection1.ImageSource.Images.SetKeyName(36, "0036.png");
+      this.sharedImageCollection1.ImageSource.Images.SetKeyName(37, "0037.png");
       this.sharedImageCollection1.ParentControl = null;
       ((System.ComponentModel.ISupportInitialize)(this.sharedImageCollection1.ImageSource)).EndInit();
       ((System.ComponentModel.ISupportInitialize)(this.sharedImageCollection1)).EndInit();